+++ CHANGELOG.md
@@ -6,7 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+
+### Fixed
+
+- The `Running command` log message now prints immediately before the child
+  process is spawned. Previously it would print even if it was blocked by
+  parallelism contention.
 
 ## [0.4.0] - 2022-05-06